added ability for clickable links
hyperlinks can now be clicked

to be implemented by users of the libary using a delegate that they adhere to, that way iOS and macOS can each trigger links appropriately.

in order to support this, loading the viewer and the editor html files now happens slightly different for each, this is limitation of webkit.